LayoutTests/svg/animations/svgenum-animation-9-expected.txt

SVG 1.1 dynamic animation tests

test
Test SVGTextPathSpacingType/SVGTextPathMethodType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textPath.method.animVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.method.baseVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.spacing.animVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.spacing.baseVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.method.animVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.method.baseVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.spacing.animVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.spacing.baseVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.method.animVal is SVGTextPathElement.TEXTPATH_METHODTYPE_STRETCH
PASS textPath.method.baseVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.spacing.animVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_EXACT
PASS textPath.spacing.baseVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.method.animVal is SVGTextPathElement.TEXTPATH_METHODTYPE_STRETCH
PASS textPath.method.baseVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.spacing.animVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_EXACT
PASS textPath.spacing.baseVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS textPath.method.animVal is SVGTextPathElement.TEXTPATH_METHODTYPE_STRETCH
PASS textPath.method.baseVal is SVGTextPathElement.TEXTPATH_METHODTYPE_ALIGN
PASS textPath.spacing.animVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_EXACT
PASS textPath.spacing.baseVal is SVGTextPathElement.TEXTPATH_SPACINGTYPE_AUTO
PASS successfullyParsed is true

TEST COMPLETE
